+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Ignore outdated/removed artifact mods still attached to armor.
 * You can now select Stasis subclasses in the Loadout Optimizer and use the stat effects from fragments.
 * When determining mod assignments, DIM will now consider in game mod placement and attempt to use the same position if possible.
+* Tracked Triumphs are now grouped together with similar records.
 
 ## 6.97.1 <span class="changelog-date">(2021-12-26)</span>